+++ tests/Test/DirectiveSpec.hs
@@ -20,7 +20,7 @@ experimentalResolver = Schema { query = queryType, mutation = Nothing }
     resolver = pure $ Int 5
     queryType = Out.ObjectType "Query" Nothing []
         $ HashMap.singleton "experimentalField"
-        $ Out.Field Nothing (Out.NamedScalarType int) mempty resolver
+        $ Out.Resolver (Out.Field Nothing (Out.NamedScalarType int) mempty) resolver
 
 emptyObject :: Aeson.Value
 emptyObject = object